+++ src/packages/piral-mithril/src/create.ts
@@ -1,5 +1,6 @@
 import * as m from 'mithril';
 import { Extend } from 'piral-core';
+import { createConverter } from './converter';
 import { PiletMithrilApi } from './types';
 
 /**
@@ -20,32 +21,8 @@ export function createMithrilApi(config: MithrilConfig = {}): Extend<PiletMithri
   const { rootName = 'slot' } = config;
 
   return context => {
-    context.converters.mithril = ({ component, captured }) => ({
-      mount(el, props, ctx) {
-        m.mount(el, {
-          view: () =>
-            m(component, {
-              ...captured,
-              ...ctx,
-              ...props,
-            }),
-        });
-      },
-      update(el, props, ctx) {
-        m.mount(el, {
-          view: () =>
-            m(component, {
-              ...captured,
-              ...ctx,
-              ...props,
-            }),
-        });
-      },
-      unmount(el) {
-        // tslint:disable-next-line:no-null-keyword
-        m.mount(el, null);
-      },
-    });
+    const convert = createConverter();
+    context.converters.mithril = ({ component, captured }) => convert(component, captured);
 
     return api => ({
       fromMithril(component, captured) {
